Extracts parts of code from submodel_database.go
